+++ blender-qmm/mats/base/Tin.py
@@ -47,6 +47,7 @@ class QMMTin(bpy.types.Operator):
 
         # princibledbsdf
         BSDF = nodes.get('Principled BSDF')
+        BSDF.distribution = 'MULTI_GGX'
         BSDF.location = (-300, 0)
         BSDF.inputs[0].default_value = (0.8, 0.8, 0.8, 1)
         BSDF.inputs[6].default_value = 1