+++ blender-qmm/mats/extras/Cinnabar.py
@@ -47,6 +47,7 @@ class QMMCinnabar(bpy.types.Operator):
 
         # princibledbsdf
         BSDF = nodes.get('Principled BSDF')
+        BSDF.distribution = 'MULTI_GGX'
         BSDF.location = (-300, 0)
         BSDF.inputs[0].default_value = (0.768151, 0.054480, 0.034340, 1)
         BSDF.inputs[6].default_value = 1