+++ blender-qmm/mats/noble/Gold.py
@@ -45,6 +45,7 @@ class QMMGold(bpy.types.Operator):
 
         # principledbsdf
         BSDF = nodes.get('Principled BSDF')
+        BSDF.distribution = 'MULTI_GGX'
         BSDF.location = (-300, 0)
         BSDF.inputs[0].default_value = (0.947306, 0.775822, 0.371238, 1)
         BSDF.inputs[6].default_value = 1