+++ blender-qmm/mats/noble/Silver.py
@@ -45,6 +45,7 @@ class QMMSilver(bpy.types.Operator):
 
         # princibledbsdf
         BSDF = nodes.get('Principled BSDF')
+        BSDF.distribution = 'MULTI_GGX'
         BSDF.location = (-300, 0)
         # BSDF.inputs[0].default_value = (0.401978, 0.396755, 0.417885, 1)
         BSDF.inputs[0].default_value = (0.962, 0.949, 0.922, 1)